Datasets
Updated tagging, but requires consideration and is not set. This is what can be done with JATS - is a related "article" not "object" as Datasets should be given the prominence of an article level and not object level. However, outcomes of the data citation group may have something better to suggest